backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
sqlparse==0.4.4
djangorestframework
djangorestframework-simplejwt
psycopg
pandas
openpyxl
drf-yasg
PyJWT
python-dotenv
pillow